Adding a solver directory correcting HALS for sparse and ridge regularizations Adding balancing Adding Tucker normalization in class Updating API, with todos
