+++ tensorly/tucker_tensor.py
@@ -305,6 +305,11 @@ class TuckerTensor(FactorizedTensor):
         return tucker_mode_dot(
             self, matrix_or_vector, mode, keep_dim=keep_dim, copy=copy
         )
+    
+    def normalize(self):
+        """TODO doc
+        """
+        self.core, self.factors = tucker_normalize(self)
 
 
 def _tucker_n_param(tensor_shape, rank):
